refactor(dca/step1): refactor step1 to be split across multiple components
