+++ src/pages/dca/create-position/components/swap/index.tsx
@@ -2,72 +2,50 @@ import React from 'react';
 import { parseUnits, formatUnits } from '@ethersproject/units';
 import Paper from '@mui/material/Paper';
 import styled from 'styled-components';
-import isUndefined from 'lodash/isUndefined';
-import { NetworkStruct, Token, YieldOption, YieldOptions } from '@types';
+import { Token, YieldOption, YieldOptions } from '@types';
 import Typography from '@mui/material/Typography';
 import Slide from '@mui/material/Slide';
 import TokenPicker from '@pages/dca/components/dca-token-picker';
-import { FormattedMessage, useIntl } from 'react-intl';
-import Button from '@common/components/button';
-import Tooltip from '@mui/material/Tooltip';
+import { FormattedMessage } from 'react-intl';
 import find from 'lodash/find';
 import useBalance from '@hooks/useBalance';
 import useUsedTokens from '@hooks/useUsedTokens';
 import StalePairModal from '@pages/dca/components/stale-pair-modal';
-import LowLiquidityModal from '@pages/dca/components/low-liquidity-modal';
-import AllowanceSplitButton from '@common/components/allowance-split-button';
 import {
-  FULL_DEPOSIT_TYPE,
-  MODE_TYPES,
   POSSIBLE_ACTIONS,
-  RATE_TYPE,
-  SUPPORTED_NETWORKS,
   TRANSACTION_TYPES,
   WHALE_MODE_FREQUENCIES,
-  WHALE_MINIMUM_VALUES,
-  TESTNETS,
   NETWORKS,
-  MAX_UINT_32,
   LATEST_VERSION,
   MINIMUM_USD_RATE_FOR_YIELD,
   DEFAULT_MINIMUM_USD_RATE_FOR_YIELD,
-  MINIMUM_USD_RATE_FOR_DEPOSIT,
-  DEFAULT_MINIMUM_USD_RATE_FOR_DEPOSIT,
-  STRING_SWAP_INTERVALS,
+  ONE_DAY,
+  shouldEnableFrequency,
+  ModeTypesIds,
 } from '@constants';
-import HelpOutlineIcon from '@mui/icons-material/HelpOutline';
 import useTransactionModal from '@hooks/useTransactionModal';
-import { emptyTokenWithAddress, parseUsdPrice, formatCurrencyAmount, usdPriceToToken } from '@common/utils/currency';
-import {
-  useTransactionAdder,
-  useHasPendingApproval,
-  useHasPendingPairCreation,
-  useHasConfirmedApproval,
-} from '@state/transactions/hooks';
+import { emptyTokenWithAddress, parseUsdPrice } from '@common/utils/currency';
+import { useTransactionAdder } from '@state/transactions/hooks';
 import { calculateStale, STALE } from '@common/utils/parsing';
 import useAvailablePairs from '@hooks/useAvailablePairs';
 import { BigNumber } from 'ethers';
-import { PROTOCOL_TOKEN_ADDRESS, getWrappedProtocolToken, EMPTY_TOKEN } from '@common/mocks/tokens';
-import CenteredLoadingIndicator from '@common/components/centered-loading-indicator';
-import useAllowance from '@hooks/useAllowance';
-import useCanSupportPair from '@hooks/useCanSupportPair';
+import { PROTOCOL_TOKEN_ADDRESS, getWrappedProtocolToken } from '@common/mocks/tokens';
 import useWalletService from '@hooks/useWalletService';
 import useContractService from '@hooks/useContractService';
 import usePositionService from '@hooks/usePositionService';
 import useRawUsdPrice from '@hooks/useUsdRawPrice';
-import useWeb3Service from '@hooks/useWeb3Service';
 import useErrorService from '@hooks/useErrorService';
-import useCurrentNetwork from '@hooks/useCurrentNetwork';
 import { shouldTrackError } from '@common/utils/errors';
 import useTrackEvent from '@hooks/useTrackEvent';
 import useReplaceHistory from '@hooks/useReplaceHistory';
 import useLoadedAsSafeApp from '@hooks/useLoadedAsSafeApp';
 import { TransactionResponse } from '@ethersproject/providers';
-import { useConnectModal } from '@rainbow-me/rainbowkit';
 import { useAppDispatch } from '@state/hooks';
-import { setNetwork } from '@state/config/actions';
+import { setFromValue, setFrom, setFrequencyType, setTo, setModeType, setRate } from '@state/create-position/actions';
+import { useCreatePositionState } from '@state/create-position/hooks';
 import SwapFirstStep from '../step1';
 import SwapSecondStep from '../step2';
+import DcaButton from '../dca-button';
 
 const StyledPaper = styled(Paper)`
   padding: 16px;
@@ -79,15 +57,6 @@ const StyledPaper = styled(Paper)`
   backdrop-filter: blur(6px);
 `;
 
-const StyledHelpOutlineIcon = styled(HelpOutlineIcon)`
-  margin-left: 10px;
-`;
-
-const StyledButton = styled(Button)`
-  padding: 10px 18px;
-  border-radius: 12px;
-`;
-
 interface AvailableSwapInterval {
   label: {
     singular: string;
@@ -97,84 +66,51 @@ interface AvailableSwapInterval {
 }
 
 interface SwapProps {
-  from: Token | null;
-  fromValue: string;
-  to: Token | null;
-  frequencyType: BigNumber;
-  frequencyValue: string;
-  setFrom: (from: Token) => void;
-  setTo: (to: Token) => void;
-  toggleFromTo: () => void;
-  setFromValue: (newFromValue: string) => void;
-  setFrequencyType: (newFrequencyType: BigNumber) => void;
   setFrequencyValue: (newFrequencyValue: string) => void;
   setYieldEnabled: (newYieldEnabled: boolean) => void;
   currentNetwork: { chainId: number; name: string };
   availableFrequencies: AvailableSwapInterval[];
-  yieldEnabled: boolean;
   yieldOptions: YieldOptions;
   isLoadingYieldOptions: boolean;
-  fromYield: YieldOption | null | undefined;
-  toYield: YieldOption | null | undefined;
   setFromYield: (newYield?: null | YieldOption) => void;
   setToYield: (newYield?: null | YieldOption) => void;
   handleChangeNetwork: (newChainId: number) => void;
 }
 
 const Swap = ({
-  from,
-  to,
-  fromValue,
-  setFrom,
-  setTo,
-  toggleFromTo,
-  setFromValue,
-  setFrequencyType,
   setFrequencyValue,
-  frequencyType,
-  frequencyValue,
   currentNetwork,
   availableFrequencies,
-  yieldEnabled,
   setYieldEnabled,
   yieldOptions,
   isLoadingYieldOptions,
-  fromYield,
-  toYield,
   setFromYield,
   setToYield,
   handleChangeNetwork,
 }: SwapProps) => {
-  const web3Service = useWeb3Service();
+  const { fromValue, frequencyType, frequencyValue, from, to, yieldEnabled, fromYield, toYield, modeType, rate } =
+    useCreatePositionState();
   const containerRef = React.useRef(null);
   const [createStep, setCreateStep] = React.useState<0 | 1>(0);
   const [showFirstStep, setShowFirstStep] = React.useState(false);
   const [showSecondStep, setShowSecondStep] = React.useState(false);
   const [isRender, setIsRender] = React.useState(true);
-  const { openConnectModal } = useConnectModal();
-  const [modeType, setModeType] = React.useState(MODE_TYPES.FULL_DEPOSIT.id);
-  const [rate, setRate] = React.useState('0');
   const [shouldShowPicker, setShouldShowPicker] = React.useState(false);
   const [selecting, setSelecting] = React.useState(from || emptyTokenWithAddress('from'));
   const [shouldShowStalePairModal, setShouldShowStalePairModal] = React.useState(false);
-  const [shouldShowLowLiquidityModal, setShouldShowLowLiquidityModal] = React.useState(false);
   const [currentAction, setCurrentAction] = React.useState<keyof typeof POSSIBLE_ACTIONS>('createPosition');
-  const [isLoading] = React.useState(false);
-  const [whaleMode, setWhaleMode] = React.useState(false);
+  const [whaleMode] = React.useState(false);
   const [setModalSuccess, setModalLoading, setModalError] = useTransactionModal();
   const addTransaction = useTransactionAdder();
   const walletService = useWalletService();
   const positionService = usePositionService();
   const dispatch = useAppDispatch();
   const contractService = useContractService();
-  const intl = useIntl();
   const availablePairs = useAvailablePairs();
   const errorService = useErrorService();
   const trackEvent = useTrackEvent();
   // const pairService = usePairService();
   const [balance, , balanceErrors] = useBalance(from);
-  const actualCurrentNetwork = useCurrentNetwork();
-  const isOnCorrectNetwork = actualCurrentNetwork.chainId === currentNetwork.chainId;
   const [usedTokens] = useUsedTokens();
 
   const existingPair = React.useMemo(() => {
@@ -197,15 +133,7 @@ const Swap = ({
       (pair) => pair.token0.address === token0.toLocaleLowerCase() && pair.token1.address === token1.toLocaleLowerCase()
     );
   }, [from, to, availablePairs, (availablePairs && availablePairs.length) || 0, fromYield, toYield]);
-  const isCreatingPair = useHasPendingPairCreation(from, to);
   const loadedAsSafeApp = useLoadedAsSafeApp();
-
-  const hasPendingApproval = useHasPendingApproval(from, web3Service.getAccount(), !!fromYield?.tokenAddress);
-  const hasConfirmedApproval = useHasConfirmedApproval(from, web3Service.getAccount(), !!fromYield?.tokenAddress);
-
-  const [allowance, , allowanceErrors] = useAllowance(from, !!fromYield?.tokenAddress);
-
-  const [pairIsSupported, isLoadingPairIsSupported] = useCanSupportPair(from, to);
 
   const [usdPrice, isLoadingUsdPrice] = useRawUsdPrice(from);
 
@@ -226,14 +154,16 @@ const Swap = ({
 
   React.useEffect(() => {
     if (!from) return;
-    setRate(
-      (fromValue &&
-        parseUnits(fromValue, from.decimals).gt(BigNumber.from(0)) &&
-        frequencyValue &&
-        BigNumber.from(frequencyValue).gt(BigNumber.from(0)) &&
-        from &&
-        formatUnits(parseUnits(fromValue, from.decimals).div(BigNumber.from(frequencyValue)), from.decimals)) ||
-        '0'
+    dispatch(
+      setRate(
+        (fromValue &&
+          parseUnits(fromValue, from.decimals).gt(BigNumber.from(0)) &&
+          frequencyValue &&
+          BigNumber.from(frequencyValue).gt(BigNumber.from(0)) &&
+          from &&
+          formatUnits(parseUnits(fromValue, from.decimals).div(BigNumber.from(frequencyValue)), from.decimals)) ||
+          '0'
+      )
     );
   }, [from]);
 
@@ -250,33 +180,39 @@ const Swap = ({
     !!usdPrice &&
     rateUsdPrice >= (MINIMUM_USD_RATE_FOR_YIELD[currentNetwork.chainId] || DEFAULT_MINIMUM_USD_RATE_FOR_YIELD);
 
-  const hasEnoughUsdForDeposit =
-    !!usdPrice &&
-    rateUsdPrice >= (MINIMUM_USD_RATE_FOR_DEPOSIT[currentNetwork.chainId] || DEFAULT_MINIMUM_USD_RATE_FOR_DEPOSIT);
-
   // only allowed if set for 10 days and at least 10 USD
   const shouldEnableYield = yieldEnabled && (fromCanHaveYield || toCanHaveYield) && hasEnoughUsdForYield;
 
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  const toggleWhaleMode = () => {
-    if (whaleMode) {
-      if (
-        (WHALE_MODE_FREQUENCIES[currentNetwork.chainId] || WHALE_MODE_FREQUENCIES[NETWORKS.optimism.chainId]).includes(
-          frequencyType.toString()
-        )
-      ) {
-        const firstNonFilteredResult = availableFrequencies.filter(
-          (frequency) =>
-            !(
-              WHALE_MODE_FREQUENCIES[currentNetwork.chainId] || WHALE_MODE_FREQUENCIES[NETWORKS.optimism.chainId]
-            ).includes(frequency.value.toString())
-        )[0];
-
-        setFrequencyType(firstNonFilteredResult.value);
+  const onSetFrom = (newFrom: Token) => {
+    // check for decimals
+    if (from && newFrom.decimals < from.decimals) {
+      const splitValue = /^(\d*)\.?(\d*)$/.exec(fromValue);
+      let newFromValue = fromValue;
+      if (splitValue && splitValue[2] !== '') {
+        newFromValue = `${splitValue[1]}.${splitValue[2].substring(0, newFrom.decimals)}`;
       }
+
+      dispatch(setFromValue(newFromValue));
     }
 
-    setWhaleMode(!whaleMode);
+    dispatch(setFrom(newFrom));
+
+    if (!shouldEnableFrequency(frequencyType.toString(), newFrom.address, to?.address, currentNetwork.chainId)) {
+      dispatch(setFrequencyType(ONE_DAY));
+    }
+
+    replaceHistory(`/create/${currentNetwork.chainId}/${newFrom.address}/${to?.address || ''}`);
+    trackEvent('DCA - Set from', { fromAddress: newFrom?.address, toAddress: to?.address });
+  };
+  const onSetTo = (newTo: Token) => {
+    dispatch(setTo(newTo));
+    if (!shouldEnableFrequency(frequencyType.toString(), from?.address, newTo.address, currentNetwork.chainId)) {
+      dispatch(setFrequencyType(ONE_DAY));
+    }
+    if (from) {
+      replaceHistory(`/create/${currentNetwork.chainId}/${from.address || ''}/${newTo.address}`);
+    }
+    trackEvent('DCA - Set to', { fromAddress: from?.address, toAddress: newTo?.address });
   };
 
   const handleApproveToken = async (amount?: BigNumber) => {
@@ -407,8 +343,8 @@ const Swap = ({
         ),
       });
 
-      setFromValue('');
-      setRate('0');
+      dispatch(setFromValue(''));
+      dispatch(setRate('0'));
       setToYield(undefined);
       setFromYield(undefined);
       setCreateStep(0);
@@ -498,8 +434,8 @@ const Swap = ({
         ),
       });
 
-      setFromValue('');
-      setRate('0');
+      dispatch(setFromValue(''));
+      dispatch(setRate('0'));
       setToYield(undefined);
       setFromYield(undefined);
       setCreateStep(0);
@@ -597,31 +533,35 @@ const Swap = ({
 
   const handleFromValueChange = (newFromValue: string) => {
     if (!from) return;
-    setModeType(FULL_DEPOSIT_TYPE);
-    setFromValue(newFromValue);
-    setRate(
-      (newFromValue &&
-        parseUnits(newFromValue, from.decimals).gt(BigNumber.from(0)) &&
-        frequencyValue &&
-        BigNumber.from(frequencyValue).gt(BigNumber.from(0)) &&
-        from &&
-        formatUnits(parseUnits(newFromValue, from.decimals).div(BigNumber.from(frequencyValue)), from.decimals)) ||
-        '0'
+    dispatch(setModeType(ModeTypesIds.FULL_DEPOSIT_TYPE));
+    dispatch(setFromValue(newFromValue));
+    dispatch(
+      setRate(
+        (newFromValue &&
+          parseUnits(newFromValue, from.decimals).gt(BigNumber.from(0)) &&
+          frequencyValue &&
+          BigNumber.from(frequencyValue).gt(BigNumber.from(0)) &&
+          from &&
+          formatUnits(parseUnits(newFromValue, from.decimals).div(BigNumber.from(frequencyValue)), from.decimals)) ||
+          '0'
+      )
     );
   };
 
   const handleRateValueChange = (newRate: string) => {
     if (!from) return;
-    setModeType(RATE_TYPE);
-    setRate(newRate);
-    setFromValue(
-      (newRate &&
-        parseUnits(newRate, from.decimals).gt(BigNumber.from(0)) &&
-        frequencyValue &&
-        BigNumber.from(frequencyValue).gt(BigNumber.from(0)) &&
-        from &&
-        formatUnits(parseUnits(newRate, from.decimals).mul(BigNumber.from(frequencyValue)), from.decimals)) ||
-        ''
+    dispatch(setModeType(ModeTypesIds.RATE_TYPE));
+    dispatch(setRate(newRate));
+    dispatch(
+      setFromValue(
+        (newRate &&
+          parseUnits(newRate, from.decimals).gt(BigNumber.from(0)) &&
+          frequencyValue &&
+          BigNumber.from(frequencyValue).gt(BigNumber.from(0)) &&
+          from &&
+          formatUnits(parseUnits(newRate, from.decimals).mul(BigNumber.from(frequencyValue)), from.decimals)) ||
+          ''
+      )
     );
     trackEvent('DCA - Set rate step 2');
   };
@@ -629,31 +569,31 @@ const Swap = ({
   const handleFrequencyChange = (newFrequencyValue: string) => {
     if (!from) return;
     setFrequencyValue(newFrequencyValue);
-    if (modeType === RATE_TYPE) {
-      setFromValue(
-        (rate &&
-          parseUnits(rate, from.decimals).gt(BigNumber.from(0)) &&
-          newFrequencyValue &&
-          BigNumber.from(newFrequencyValue).gt(BigNumber.from(0)) &&
-          from &&
-          formatUnits(parseUnits(rate, from.decimals).mul(BigNumber.from(newFrequencyValue)), from.decimals)) ||
-          ''
+    if (modeType === ModeTypesIds.RATE_TYPE) {
+      dispatch(
+        setFromValue(
+          (rate &&
+            parseUnits(rate, from.decimals).gt(BigNumber.from(0)) &&
+            newFrequencyValue &&
+            BigNumber.from(newFrequencyValue).gt(BigNumber.from(0)) &&
+            from &&
+            formatUnits(parseUnits(rate, from.decimals).mul(BigNumber.from(newFrequencyValue)), from.decimals)) ||
+            ''
+        )
       );
     } else {
-      setRate(
-        (fromValue &&
-          parseUnits(fromValue, from.decimals).gt(BigNumber.from(0)) &&
-          newFrequencyValue &&
-          BigNumber.from(newFrequencyValue).gt(BigNumber.from(0)) &&
-          from &&
-          formatUnits(parseUnits(fromValue, from.decimals).div(BigNumber.from(newFrequencyValue)), from.decimals)) ||
-          '0'
+      dispatch(
+        setRate(
+          (fromValue &&
+            parseUnits(fromValue, from.decimals).gt(BigNumber.from(0)) &&
+            newFrequencyValue &&
+            BigNumber.from(newFrequencyValue).gt(BigNumber.from(0)) &&
+            from &&
+            formatUnits(parseUnits(fromValue, from.decimals).div(BigNumber.from(newFrequencyValue)), from.decimals)) ||
+            '0'
+        )
       );
     }
-  };
-
-  const onLowLiquidityModalClose = () => {
-    setShouldShowLowLiquidityModal(false);
   };
 
   const POSSIBLE_ACTIONS_FUNCTIONS = {
@@ -671,7 +611,7 @@ const Swap = ({
   };
 
   // eslint-disable-next-line @typescript-eslint/require-await
-  const checkForLowLiquidity = async (actionToDo: keyof typeof POSSIBLE_ACTIONS, amount?: BigNumber) => {
+  const onButtonClick = async (actionToDo: keyof typeof POSSIBLE_ACTIONS, amount?: BigNumber) => {
     setCurrentAction(actionToDo);
     if (PRE_POSSIBLE_ACTIONS_FUNCTIONS[actionToDo]) {
       // eslint-disable-next-line @typescript-eslint/no-floating-promises
@@ -685,308 +625,9 @@ const Swap = ({
         trackEvent('DCA - Create position', { fromAddress: from?.address, toAddress: to?.address });
       }
     }
-    /**  Disable low liquidity modal for now */
-    // setIsLoading(true);
-    // if (!from || !to) return;
-
-    // const oracle = await pairService.getPairOracle(
-    //   { tokenA: fromYield?.tokenAddress || from.address, tokenB: toYield?.tokenAddress || to.address },
-    //   !!existingPair
-    // );
-
-    // let hasLowLiquidity = oracle === ORACLES.UNISWAP;
-
-    // if (oracle === ORACLES.UNISWAP) {
-    //   try {
-    //     const liquidity = await pairService.getPairLiquidity(from, to);
-    //     hasLowLiquidity = liquidity <= MINIMUM_LIQUIDITY_USD;
-    //   } catch {
-    //     hasLowLiquidity = false;
-    //   }
-    // }
-
-    // setIsLoading(false);
-
-    // setCurrentAction(actionToDo);
-    // if (hasLowLiquidity) {
-    //   setShouldShowLowLiquidityModal(true);
-    // } else if (PRE_POSSIBLE_ACTIONS_FUNCTIONS[actionToDo]) {
-    //   // eslint-disable-next-line @typescript-eslint/no-floating-promises
-    //   PRE_POSSIBLE_ACTIONS_FUNCTIONS[actionToDo]();
-    // }
   };
 
-  const closeLowLiquidityModal = () => {
-    if (PRE_POSSIBLE_ACTIONS_FUNCTIONS[currentAction]) {
-      // eslint-disable-next-line @typescript-eslint/no-floating-promises
-      PRE_POSSIBLE_ACTIONS_FUNCTIONS[currentAction]();
-    }
-    onLowLiquidityModalClose();
-  };
-
-  const cantFund = from && !!fromValue && !!balance && parseUnits(fromValue, from.decimals).gt(balance);
-
-  const isApproved =
-    !from ||
-    hasConfirmedApproval ||
-    (from &&
-      (!fromValue
-        ? true
-        : (allowance.allowance &&
-            allowance.token.address === from.address &&
-            parseUnits(allowance.allowance, from.decimals).gte(parseUnits(fromValue, from.decimals))) ||
-          from.address === PROTOCOL_TOKEN_ADDRESS));
-
-  const shouldDisableApproveButton =
-    !from ||
-    !to ||
-    !fromValue ||
-    !frequencyValue ||
-    cantFund ||
-    !balance ||
-    balanceErrors ||
-    allowanceErrors ||
-    parseUnits(fromValue, from.decimals).lte(BigNumber.from(0)) ||
-    BigNumber.from(frequencyValue).lte(BigNumber.from(0)) ||
-    (shouldEnableYield && fromCanHaveYield && isUndefined(fromYield)) ||
-    (shouldEnableYield && toCanHaveYield && isUndefined(toYield));
-
-  const onChangeNetwork = (chainId: number) => {
-    // eslint-disable-next-line @typescript-eslint/no-floating-promises
-    walletService.changeNetwork(chainId, () => {
-      const networkToSet = find(NETWORKS, { chainId });
-      replaceHistory(`/create/${chainId}`);
-      dispatch(setNetwork(networkToSet as NetworkStruct));
-      if (networkToSet) {
-        web3Service.setNetwork(networkToSet?.chainId);
-      }
-    });
-    trackEvent('DCA - Change network', { chainId });
-  };
-
-  const shouldDisableButton = shouldDisableApproveButton || !isApproved;
-
-  const isTestnet = TESTNETS.includes(currentNetwork.chainId);
-
-  let shouldShowNotEnoughForWhale =
-    from &&
-    whaleMode &&
-    (WHALE_MODE_FREQUENCIES[currentNetwork.chainId] || WHALE_MODE_FREQUENCIES[NETWORKS.optimism.chainId]).includes(
-      frequencyType.toString()
-    ) &&
-    fromValue &&
-    frequencyValue &&
-    !isLoadingUsdPrice &&
-    usdPrice &&
-    parseFloat(formatUnits(parseUnits(fromValue, from.decimals).mul(BigNumber.from(frequencyValue)), from.decimals)) *
-      fromValueUsdPrice <
-      (WHALE_MINIMUM_VALUES[currentNetwork.chainId][frequencyType.toString()] || Infinity);
-
-  shouldShowNotEnoughForWhale =
-    !isTestnet &&
-    (shouldShowNotEnoughForWhale ||
-      (whaleMode &&
-        !usdPrice &&
-        (WHALE_MODE_FREQUENCIES[currentNetwork.chainId] || WHALE_MODE_FREQUENCIES[NETWORKS.optimism.chainId]).includes(
-          frequencyType.toString()
-        )));
-  const NotConnectedButton = (
-    <StyledButton size="large" variant="contained" fullWidth color="error">
-      <Typography variant="body1">
-        <FormattedMessage description="wrong chainId" defaultMessage="We do not support this chain yet" />
-      </Typography>
-    </StyledButton>
-  );
-
-  const NoWalletButton = (
-    <StyledButton size="large" color="primary" variant="contained" fullWidth onClick={openConnectModal}>
-      <Typography variant="body1">
-        <FormattedMessage description="connect wallet" defaultMessage="Connect wallet" />
-      </Typography>
-    </StyledButton>
-  );
-
-  const IncorrectNetworkButton = (
-    <StyledButton
-      size="large"
-      color="secondary"
-      variant="contained"
-      onClick={() => onChangeNetwork(currentNetwork.chainId)}
-      fullWidth
-    >
-      <Typography variant="body1">
-        <FormattedMessage
-          description="incorrect network"
-          defaultMessage="Change network to {network}"
-          values={{ network: currentNetwork.name }}
-        />
-      </Typography>
-    </StyledButton>
-  );
-
-  const isApproveTokenDisabled = !!isApproved || hasPendingApproval || isLoading || !!shouldDisableApproveButton;
-
-  const ApproveTokenButton = (
-    <AllowanceSplitButton
-      onMaxApprove={() => checkForLowLiquidity(POSSIBLE_ACTIONS.approveToken as keyof typeof POSSIBLE_ACTIONS)}
-      onApproveExact={(amount) =>
-        checkForLowLiquidity(POSSIBLE_ACTIONS.approveTokenExact as keyof typeof POSSIBLE_ACTIONS, amount)
-      }
-      amount={from && (fromValue ? parseUnits(fromValue, from?.decimals) : null)}
-      disabled={isApproveTokenDisabled}
-      token={from}
-      tokenYield={fromYield}
-    />
-  );
-
-  const swapsIsMax = BigNumber.from(frequencyValue || '0').gt(BigNumber.from(MAX_UINT_32));
-
-  const StartPositionButton = (
-    <StyledButton
-      size="large"
-      variant="contained"
-      disabled={
-        !!shouldDisableButton || isLoading || isLoadingPairIsSupported || !!shouldShowNotEnoughForWhale || swapsIsMax
-      }
-      color="secondary"
-      fullWidth
-      onClick={() => checkForLowLiquidity(POSSIBLE_ACTIONS.createPosition as keyof typeof POSSIBLE_ACTIONS)}
-    >
-      {!isLoading && !isLoadingPairIsSupported && !isLoadingUsdPrice && !shouldShowNotEnoughForWhale && swapsIsMax && (
-        <Typography variant="body1">
-          <FormattedMessage
-            description="swapsCannotBeMax"
-            defaultMessage="Amount of swaps cannot be higher than {MAX_UINT_32}"
-            values={{ MAX_UINT_32 }}
-          />
-        </Typography>
-      )}
-      {!isLoading && !isLoadingPairIsSupported && !isLoadingUsdPrice && !shouldShowNotEnoughForWhale && !swapsIsMax && (
-        <Typography variant="body1">
-          <FormattedMessage description="create position" defaultMessage="Create position" />
-        </Typography>
-      )}
-      {!isLoading && !isLoadingPairIsSupported && !isLoadingUsdPrice && shouldShowNotEnoughForWhale && !swapsIsMax && (
-        <Typography variant="body1">
-          <FormattedMessage
-            description="notenoughwhale"
-            defaultMessage="You can only deposit with a minimum value of {value} USD"
-            values={{ value: WHALE_MINIMUM_VALUES[currentNetwork.chainId][frequencyType.toString()] }}
-          />
-        </Typography>
-      )}
-      {(isLoading || isLoadingPairIsSupported || isLoadingUsdPrice) && <CenteredLoadingIndicator />}
-    </StyledButton>
-  );
-
-  const SafeApproveAndStartPositionButton = (
-    <StyledButton
-      size="large"
-      variant="contained"
-      disabled={
-        !!shouldDisableApproveButton ||
-        isLoading ||
-        isLoadingPairIsSupported ||
-        !!shouldShowNotEnoughForWhale ||
-        swapsIsMax
-      }
-      color="secondary"
-      fullWidth
-      onClick={() => checkForLowLiquidity(POSSIBLE_ACTIONS.approveAndCreatePosition as keyof typeof POSSIBLE_ACTIONS)}
-    >
-      {!isLoading && !isLoadingPairIsSupported && !isLoadingUsdPrice && !shouldShowNotEnoughForWhale && swapsIsMax && (
-        <Typography variant="body1">
-          <FormattedMessage
-            description="swapsCannotBeMax"
-            defaultMessage="Amount of swaps cannot be higher than {MAX_UINT_32}"
-            values={{ MAX_UINT_32 }}
-          />
-        </Typography>
-      )}
-      {!isLoading && !isLoadingPairIsSupported && !isLoadingUsdPrice && !shouldShowNotEnoughForWhale && !swapsIsMax && (
-        <Typography variant="body1">
-          <FormattedMessage
-            description="approve and create position"
-            defaultMessage="Approve {from} and create position"
-            values={{ from: from?.symbol || '' }}
-          />
-        </Typography>
-      )}
-      {!isLoading && !isLoadingPairIsSupported && !isLoadingUsdPrice && shouldShowNotEnoughForWhale && !swapsIsMax && (
-        <Typography variant="body1">
-          <FormattedMessage
-            description="notenoughwhale"
-            defaultMessage="You can only deposit with a minimum value of {value} USD"
-            values={{ value: WHALE_MINIMUM_VALUES[currentNetwork.chainId][frequencyType.toString()] }}
-          />
-        </Typography>
-      )}
-      {(isLoading || isLoadingPairIsSupported || isLoadingUsdPrice) && <CenteredLoadingIndicator />}
-    </StyledButton>
-  );
-
-  const CreatingPairButton = (
-    <StyledButton size="large" variant="contained" disabled color="secondary" fullWidth>
-      <Typography variant="body1">
-        <FormattedMessage description="creating pair" defaultMessage="Creating this pair" />
-      </Typography>
-    </StyledButton>
-  );
-
-  const NoFundsButton = (
-    <StyledButton size="large" color="default" variant="contained" fullWidth disabled>
-      <Typography variant="body1">
-        <FormattedMessage description="insufficient funds" defaultMessage="Insufficient funds" />
-      </Typography>
-    </StyledButton>
-  );
-
-  const minimumTokensNeeded = usdPriceToToken(
-    from,
-    MINIMUM_USD_RATE_FOR_DEPOSIT[currentNetwork.chainId] || DEFAULT_MINIMUM_USD_RATE_FOR_DEPOSIT,
-    usdPrice
-  );
-
-  const NoMinForDepositButton = (
-    <StyledButton size="large" color="default" variant="contained" fullWidth disabled>
-      <Typography variant="body1">
-        <FormattedMessage
-          description="disabledDepositByUsdValue"
-          // eslint-disable-next-line no-template-curly-in-string
-          defaultMessage="The position must have a minimum rate of ${minimum} USD ({minToken} {symbol}) per {frequency} to be created."
-          values={{
-            minimum: MINIMUM_USD_RATE_FOR_DEPOSIT[currentNetwork.chainId] || MINIMUM_USD_RATE_FOR_DEPOSIT,
-            minToken: formatCurrencyAmount(minimumTokensNeeded, from || EMPTY_TOKEN, 3, 3),
-            symbol: from?.symbol || '',
-            frequency: intl.formatMessage(
-              STRING_SWAP_INTERVALS[frequencyType.toString() as keyof typeof STRING_SWAP_INTERVALS].singularSubject
-            ),
-          }}
-        />
-      </Typography>
-    </StyledButton>
-  );
-
-  const PairNotSupportedButton = (
-    <StyledButton size="large" color="error" variant="contained" fullWidth disabled style={{ pointerEvents: 'all' }}>
-      <Typography variant="body1">
-        <FormattedMessage description="pairNotOnUniswap" defaultMessage="We do not support this pair" />
-      </Typography>
-      <Tooltip
-        title="If you want to use this pair of tokens you must first create a pool for it on UniswapV3"
-        arrow
-        placement="top"
-      >
-        <StyledHelpOutlineIcon fontSize="small" />
-      </Tooltip>
-    </StyledButton>
-  );
-
-  const LoadingButton = (
-    <StyledButton size="large" color="default" variant="contained" fullWidth disabled>
-      <CenteredLoadingIndicator />
-    </StyledButton>
-  );
+  const cantFund = !!from && !!fromValue && !!balance && parseUnits(fromValue, from.decimals).gt(balance);
 
   const handleSetStep = (step: 0 | 1) => {
     if (isRender) {
@@ -1001,48 +642,23 @@ const Swap = ({
     }
   };
 
-  const NextStepButton = (
-    <StyledButton
-      size="large"
-      disabled={!from || !to || !fromValue || parseFloat(fromValue) === 0 || !frequencyValue || frequencyValue === '0'}
-      color="secondary"
-      variant="contained"
-      fullWidth
-      onClick={() => handleSetStep(1)}
-    >
-      <Typography variant="body1">
-        <FormattedMessage description="continue" defaultMessage="Continue" />
-      </Typography>
-    </StyledButton>
+  const ButtonToShow = (
+    <DcaButton
+      onClick={onButtonClick}
+      cantFund={cantFund}
+      usdPrice={usdPrice}
+      shouldEnableYield={shouldEnableYield}
+      balance={balance}
+      balanceErrors={balanceErrors}
+      fromCanHaveYield={fromCanHaveYield}
+      toCanHaveYield={toCanHaveYield}
+      isLoadingUsdPrice={isLoadingUsdPrice}
+      handleSetStep={handleSetStep}
+      step={createStep}
+      rateUsdPrice={rateUsdPrice}
+      fromValueUsdPrice={fromValueUsdPrice}
+    />
   );
-
-  let ButtonToShow;
-
-  if (!web3Service.getAccount()) {
-    ButtonToShow = NoWalletButton;
-  } else if (!SUPPORTED_NETWORKS.includes(currentNetwork.chainId)) {
-    ButtonToShow = NotConnectedButton;
-  } else if (isLoading || isLoadingPairIsSupported) {
-    ButtonToShow = LoadingButton;
-  } else if (!isOnCorrectNetwork) {
-    ButtonToShow = IncorrectNetworkButton;
-  } else if (!pairIsSupported && !isLoadingPairIsSupported && from && to) {
-    ButtonToShow = PairNotSupportedButton;
-  } else if (cantFund) {
-    ButtonToShow = NoFundsButton;
-  } else if (!createStep) {
-    ButtonToShow = NextStepButton;
-  } else if (!hasEnoughUsdForDeposit) {
-    ButtonToShow = NoMinForDepositButton;
-  } else if (!isApproved && balance && balance.gt(BigNumber.from(0)) && to && loadedAsSafeApp) {
-    ButtonToShow = SafeApproveAndStartPositionButton;
-  } else if (!isApproved && balance && balance.gt(BigNumber.from(0)) && to) {
-    ButtonToShow = ApproveTokenButton;
-  } else if (isCreatingPair) {
-    ButtonToShow = CreatingPairButton;
-  } else {
-    ButtonToShow = StartPositionButton;
-  }
 
   const filteredFrequencies = whaleMode
     ? availableFrequencies
@@ -1065,18 +681,12 @@ const Swap = ({
         }}
         onCancel={() => setShouldShowStalePairModal(false)}
       />
-      <LowLiquidityModal
-        open={shouldShowLowLiquidityModal}
-        onConfirm={closeLowLiquidityModal}
-        onCancel={onLowLiquidityModalClose}
-        actionToTake={currentAction}
-      />
 
       <TokenPicker
         shouldShow={shouldShowPicker}
         onClose={() => setShouldShowPicker(false)}
         isFrom={selecting === from}
-        onChange={(from && selecting.address === from.address) || selecting.address === 'from' ? setFrom : setTo}
+        onChange={(from && selecting.address === from.address) || selecting.address === 'from' ? onSetFrom : onSetTo}
         usedTokens={usedTokens}
         ignoreValues={[]}
         yieldOptions={yieldOptions}
@@ -1099,16 +709,8 @@ const Swap = ({
         easing="ease-out"
       >
         <SwapFirstStep
-          from={from}
-          to={to}
-          fromValue={fromValue}
-          toggleFromTo={toggleFromTo}
-          setFrequencyType={setFrequencyType}
-          frequencyType={frequencyType}
-          frequencyValue={frequencyValue}
           startSelectingCoin={startSelectingCoin}
           cantFund={cantFund}
-          handleFromValueChange={handleFromValueChange}
           balance={balance}
           frequencies={filteredFrequencies}
           handleFrequencyChange={handleFrequencyChange}
@@ -1116,6 +718,7 @@ const Swap = ({
           show={showFirstStep}
           fromValueUsdPrice={fromValueUsdPrice}
           onChangeNetwork={handleChangeNetwork}
+          handleFromValueChange={handleFromValueChange}
         />
       </Slide>
       <Slide
