+++ src/state/create-position/actions.ts
@@ -1,6 +1,7 @@
 import { createAction } from '@reduxjs/toolkit';
 import { BigNumber } from 'ethers';
 import { Token, YieldOption } from '@types';
+import { ModeTypesIds } from '@constants';
 
 export const setFromValue = createAction<string>('createPosition/setFromValue');
 
@@ -17,5 +18,9 @@ export const setYieldEnabled = createAction<boolean>('createPosition/setYieldEna
 export const setFromYield = createAction<YieldOption | null | undefined>('createPosition/setFromYield');
 
 export const setToYield = createAction<YieldOption | null | undefined>('createPosition/setToYield');
+
+export const setRate = createAction<string>('createPosition/setRate');
+
+export const setModeType = createAction<ModeTypesIds>('createPosition/setModeType');
 
 export const setDCAChainId = createAction<number>('createPosition/setDCAChainId');
